# Refactored metrics processing for metrics-agent plugin to be more tes…