…table

* Broke out `processMessages()` into separate functions to make things more testable and better suited for benchmarking.
* Added unit tests
* Added benchmarks  (there is a very slight improvement in performance compared to the monolithic `processMessages`)

**Before:**
BenchmarkMacAddrNetworkMapNoLock | 100000000 | 11.7 ns/op
BenchmarkMacAddrNetworkMap | 100000000 | 11.8 ns/op
BenchmarkCloudInstances | 30000000 | 37.2 ns/op
BenchmarkMetricsAVSVswitchIfc | 100000 | 12189 ns/op
BenchmarkMetricsAVSVswitchPort | 100000 | 13264 ns/op
BenchmarkMetricsAVSVswitchPortQueue | 200000 | 6476 ns/op
BenchmarkMetricsHostProc | 50000 | 22329 ns/op
BenchmarkMetricsIntelPCMCore | 100000 | 11045 ns/op
BenchmarkMetricsIntelRDTCore | 500000 | 3222 ns/op
BenchmarkMetricsLibvirtDomain | 100000 | 11200 ns/op
BenchmarkMetricsLibvirtDomainBlock | 200000 | 6430 ns/op
BenchmarkMetricsLibvirtDomainCore | 300000 | 4256 ns/op
BenchmarkMetricsLibvirtDomainIfc | 100000 | 12600 ns/op
BenchmarkMetricsVswitchDpdkIfc | 50000 | 24235 ns/op
BenchmarkMetricsVswitchIfc | 100000 | 15403 ns/op
BenchmarkHasSupportedMeasurementHostProcMetrics | 300000000 | 4.99 ns/op
BenchmarkHasSupportedMeasurementAvsVswitchPortQueueMetrics | 200000000 | 5.74 ns/op

**After:**
BenchmarkMacAddrNetworkMapNoLock | 100000000 | 11.5 ns/op
BenchmarkMacAddrNetworkMap | 100000000 | 13.3 ns/op
BenchmarkCloudInstances | 50000000 | 39.3 ns/op
BenchmarkMetricsAVSVswitchIfc | 100000 | 12144 ns/op
BenchmarkMetricsAVSVswitchPort | 100000 | 12279 ns/op
BenchmarkMetricsAVSVswitchPortQueue | 200000 | 5113 ns/op
BenchmarkMetricsHostProc | 50000 | 20871 ns/op
BenchmarkMetricsIntelPCMCore | 100000 | 13845 ns/op
BenchmarkMetricsIntelRDTCore | 500000 | 3148 ns/op
BenchmarkMetricsLibvirtDomain | 200000 | 11544 ns/op
BenchmarkMetricsLibvirtDomainBlock | 200000 | 6771 ns/op
BenchmarkMetricsLibvirtDomainCore | 500000 | 2792 ns/op
BenchmarkMetricsLibvirtDomainIfc | 100000 | 11530 ns/op
BenchmarkMetricsVswitchDpdkIfc | 100000 | 23302 ns/op
BenchmarkMetricsVswitchIfc | 100000 | 13374 ns/op
BenchmarkHasSupportedMeasurementHostProcMetrics | 300000000 | 5.32 ns/op
BenchmarkHasSupportedMeasurementAvsVswitchPortQueueMetrics | 200000000 | 5.95 ns/op

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/spirentorion/telegraf/16)
<!-- Reviewable:end -->