@rmyers808 2018-05-22
Review status: 0 of 2 files reviewed at latest revision, all discussions resolved.

---

*[plugins/inputs/host_agent_consumer/host_agent_consumer.go, line 255 at r1](https://reviewable.io/reviews/spirentorion/telegraf/16#-LD8kKmrE5Ex_ZGLw6cv:-LD8kKmrE5Ex_ZGLw6cw:b-g3eop5) ([raw file](https://github.com/spirentorion/telegraf/blob/03ab3657484021a9f5c4e135f609ba7fcf38014a/plugins/inputs/host_agent_consumer/host_agent_consumer.go#L255)):*
> ```Go
> 			values[*v.Name] = v.GetInt64Value()
> 		default:
> 			errors.New("unexpected metric value type")
> ```

Probably okay to not panic but wondering how this error is used?

---

*[plugins/inputs/host_agent_consumer/host_agent_consumer.go, line 293 at r1](https://reviewable.io/reviews/spirentorion/telegraf/16#-LD8lBg467mT6r93rK2Q:-LD8lBg467mT6r93rK2R:b-76hy7i) ([raw file](https://github.com/spirentorion/telegraf/blob/03ab3657484021a9f5c4e135f609ba7fcf38014a/plugins/inputs/host_agent_consumer/host_agent_consumer.go#L293)):*
> ```Go
> 	}
> 	// set the Instance Name given the libvirt UUID
> 	if hasSupportedMeasurement(measurement) {
> ```

This check was a bit odd before.  It was checking for metrics that could be assigned an instance name but may not always have one.  It's missing system tables (host_metircs,  intel_pcm_system_metrics).  Wonder if support func name should be specialized for InstanceName or maybe should just check on dimension "libvirt_uuid".

---


*Comments from [Reviewable](https://reviewable.io/reviews/spirentorion/telegraf/16)*
<!-- Sent from Reviewable.io -->