@jhiga 2018-05-22
Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions.

---

*[plugins/inputs/host_agent_consumer/host_agent_consumer.go, line 293 at r1](https://reviewable.io/reviews/spirentorion/telegraf/16#-LD8lBg467mT6r93rK2Q:-LD9AljT1Cub1tQx47oX:bb6un1j) ([raw file](https://github.com/spirentorion/telegraf/blob/03ab3657484021a9f5c4e135f609ba7fcf38014a/plugins/inputs/host_agent_consumer/host_agent_consumer.go#L293)):*
<details><summary><i>Previously, jhiga wrote…</i></summary><blockquote>

according to the docs on github, I'm not seeing `instance_name` as a listed dimension for host_metrics or for intel_pcm_system_metrics.

I'm thinking that maybe we should just check on dimension "libvirt_uuid" and get rid of `hasSupportedMetrics` altogether.

</blockquote></details>

Done

---


*Comments from [Reviewable](https://reviewable.io/reviews/spirentorion/telegraf/16)*
<!-- Sent from Reviewable.io -->