@rmyers808 2018-05-23
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io/reviews/spirentorion/telegraf/16#-:-LD9RRV5EudSos7Lzsrd:bnfp4nl)*
<!-- Sent from Reviewable.io -->